pyproject.toml
[project]
name = "uv-hackathon"
version = "0.1.0"
description = "Hackathon prototype with Instagram/YouTube crawling + LangGraph"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # --- 영상/이미지 ---
    "moviepy>=2.2.1",
    "opencv-python-headless>=4.12.0.88",
    "scenedetect>=0.6.7.1",
    "ultralytics>=8.3.214",
    "yt-dlp>=2025.10.14",
    # --- 크롤링 ---
    "selenium>=4.27.1",
    "webdriver-manager>=4.0.2",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.3",
    # --- YouTube API ---
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    # --- 환경변수 관리 ---
    "python-dotenv>=1.0.1",
    # --- LangChain / LangGraph ---
    "langchain>=0.3.27",
    "langchain-core>=0.3.79",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.35",
    "langgraph>=1.0.0",
    "langgraph-checkpoint>=2.1.2",
    "langgraph-sdk>=0.2.9",
    # --- DB/임베딩 ---
    "chromadb>=1.2.0",
    "tiktoken>=0.12.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    # --- 웹/앱 프레임워크 ---
    "fastapi>=0.119.0",
    "uvicorn>=0.37.0",
    "streamlit>=1.50.0",
    # --- NLP / 분석 ---
    "nltk>=3.9.2",
    "yfinance>=0.2.66",
    "scikit-learn>=1.7.2",
    "transformers>=4.57.1",
    "sentence-transformers>=5.1.2",
    # --- 포스터 하베스터 전용 ---
    "pydantic>=2.12.3",
    "tqdm>=4.67.1",
    "rapidfuzz>=3.9.7",
    "PyMuPDF>=1.26.6",
    "pytrends>=4.9.2",
    "python-docx>=1.2.0",
    "cloudconvert>=2.1.0",
    # --- 기타 ---
    "aiohttp>=3.13.1",
    "aiohappyeyeballs>=2.6.1",
    "altair>=5.5.0",
    "beautifulsoup4>=4.14.2",
    "Flask>=3.1.2",
    "streamlit>=1.50.0",
    "python-multipart>=0.0.20",
    "replicate>=1.0.7",
    "matplotlib>=3.10.7",
    "serpapi>=0.1.5",
    "google-search-results>=2.4.2",
    "clip",
    "accelerate>=1.11.0",
    "pillow>=11.3.0",
    "urllib3<2.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }


